python_backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.15
langgraph==0.2.45
google-search-results==2.4.2
pydantic==2.9.2

